feat: Add code for reading charts using oexserverd
This commit adds functionality that lets you easily read your
original chart files directly. The starting and stopping of
oexserverd and setting up the IPC is handled by the library.

The user is provided with an istream based interface that lets you
read the chart files as if they were regular decrypted files.